DWH/init_scripts/requirements.txt
requests==2.31.0
pymssql==2.2.11
python-dotenv==1.0.0
schedule==1.2.0
pymysql==1.1.1
pymongo==4.6.1
neo4j==5.19.0
psycopg2-binary==2.9.9
supabase==2.7.4
